get_spectrum moved to utils
